Documentations and Routes added
