+++ templates/lmao.html
@@ -4,12 +4,12 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./styles.css">
+    <link rel="stylesheet" href="{{url_for('static', filename = './CSS/style.css')}}">
     <title>Document</title>
 </head>
 
 <body>
-<!-- <style>
+    <!-- <style>
     .bgimg {
         width: 100%;
         height: 400px;
@@ -24,13 +24,15 @@
     </div>
 lmao opened ok ok
 call pe aa rhe -->
-<div class="bg-image">
-</div>
+    <div class="bg-image">
+    </div>
 
-<div class="main">
-    <h1>Harshit</h1>
-    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Ea ratione numquam cumque animi voluptate reiciendis impedit assumenda doloribus dolore obcaecati alias nam consectetur reprehenderit necessitatibus ullam omnis, facere odio magnam.</p>
-</div>
+    <div class="main">
+        <h1>Harshit</h1>
+        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Ea ratione numquam cumque animi voluptate reiciendis
+            impedit assumenda doloribus dolore obcaecati alias nam consectetur reprehenderit necessitatibus ullam omnis,
+            facere odio magnam.</p>
+    </div>
 
 </body>
