+++ templates/login.html
@@ -19,7 +19,7 @@
         <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Password">
       </div>
       <br><br>
-      <button type="submit" class="btn btn-primary" >Sign in </button>
+      <button type="submit" class="btn btn-primary" >Submit</button>
     </form>
 
 </body>
