--- templates/style.css
@@ -1,144 +1,0 @@
-
-  p{
-      width: 100%;
-      height: 100%;
-       display:flex;
-      
-      justify-content: center;
-      align-items: center; 
-      align-self: center;
-      letter-spacing: 1px;
-      background-color: #0c1022;
-  }
-  .login_form_container{
-      position: relative;
-      width: 400px;
-      height: 470px;
-      max-width: 400px;
-      max-height: 470px;
-      background: #040717;
-      border-radius: 50px 5px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      overflow: hidden;
-      margin-top: 70px;
-  }
-  .login_form_container::before{
-      
-      position: absolute;
-      width: 170%;
-      height: 170%;
-      content: '';
-      background-image: conic-gradient(transparent, transparent, transparent, #ee00ff);
-      animation: rotate_border 6s linear infinite;
-  
-  }
-  .login_form_container::after{
-      
-      position: absolute;
-      width: 170%;
-      height: 170%;
-      content: '';
-      background-image: conic-gradient(transparent, transparent, transparent, #00ccff);
-      animation: rotate_border 6s linear infinite;
-      animation-delay: -3s;
-  }
-  @keyframes rotate_border{
-      0%{
-          transform: rotate(0deg);
-      }
-      100%{
-          transform: rotate(360deg);
-      }
-  }
-  .login_form{
-      position: absolute;
-      content: '';
-      background-color: #0c1022;
-      border-radius: 50px 5px;
-      inset: 5px;
-      padding: 50px 40px;
-      z-index: 10;
-      color: #00ccff;
-  
-  } 
-  h2{
-      font-size: 40px;
-      font-weight: 600;
-      text-align: center;
-  }
-  .input_group{
-      margin-top: 40px;
-      position:relative;
-      display:flex;
-      align-items: center;
-      justify-content: start;
-  }
-  .input_text{
-      width: 95%;
-      height: 30px;
-      background: transparent;
-      border: none;
-      outline: none;
-      border-bottom: 1px solid #00ccff;
-      font-size: 20px;
-      padding-left: 10px;
-      color: #00ccff;
-  
-  }
-  ::placeholder{
-      font-size: 15px;
-      color: #00ccff52;
-      letter-spacing: 1px;
-  
-  }
-  .fa{
-      font-size: 20px;
-  
-  }
-  #login_button{
-      position: relative;
-      width: 300px;
-      height: 40px;
-      transition: 1s;
-      margin-top: 70px;
-  
-  
-  }
-  #login_button a{
-      position: absolute;
-      width: 100%;
-      height: 100%;
-      text-decoration: none;
-      z-index: 10;
-      cursor:not-allowed;
-      font-size: 22px;
-      letter-spacing: 2px;
-      border: 1px solid #00ccff;
-      border-radius: 50px;
-      background-color: #0c1022;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      
-  }
-  .fotter {
-      margin-top: 30px;
-      display: flex;
-      justify-content: space-between;
-  
-  }
-  .fotter a{
-      text-decoration: none;
-      cursor: pointer;
-      font-size: 18px;
-  }
-  .glowIcon {
-      text-shadow: 0 0 10px #00ccff;
-  
-  }
-  
-  P{
-    align-content: center;
-  }
